requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
ast_comments
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
black
blake3==1.0.4
boto3
bs4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
compressed-tensors==0.9.3
cupy-cuda12x==13.4.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi-cli==0.0.7
fastapi==0.115.12
fastrlock==0.8.3
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gguf==0.16.3
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hdbscan
hf-xet==1.1.0
hiplot
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.0.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.9.0
jsonschema-specifications==2025.4.1
jsonschema==4.23.0
lark==1.2.2
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.78.1
opencv-python-headless==4.11.0.86
openpyxl
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions-ai==0.4.8
opentelemetry-semantic-conventions==0.47b0
outlines_core==0.1.26
outlines==0.1.11
packaging==25.0
pandas>=2.0.0
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
plotly
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.1
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow
pycountry==24.6.1
pydantic_core==2.33.2
pydantic==2.11.4
Pygments==2.19.1
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
radon==6.0.1
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich-toolkit==0.14.5
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn
scipy==1.15.3
seaborn
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typer==0.15.3
typing_extensions==4.13.2
typing-inspection==0.4.0
umap-learn
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
validators
vllm==0.8.5.post1
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.29.post2
xgboost==2.0.3
xgrammar==0.1.18
yarl==1.20.0
zipp==3.21.0